[NS-72] Organize VE transfer-related code
- Move all existing VE transfer-related code over to `io.sparkcyclone.data.transfer`

- Rename `io.sparkcyclone.data.serializer` to `io.sparkcyclone.data.serialization`
