+++ src/test/scala/io/sparkcyclone/data/transfer/RowCollectingTransferDescriptorUnitSpec.scala
@@ -1,4 +1,4 @@
-package io.sparkcyclone.cache
+package io.sparkcyclone.data.transfer
 
 import io.sparkcyclone.colvector.VeColVectorSource
 import io.sparkcyclone.annotations.VectorEngineTest
